paneler åpner seg i readme og expandable. ny hook for sjekk og åpning av paneler
